+++ src/IncomingRequestFactory.php
@@ -1,7 +1,7 @@
 <?php
 namespace Phly\Http;
 
-use Psr\Http\Message\RequestInterface as RequestInterface;
+use Psr\Http\Message\IncomingRequestInterface;
 use stdClass;
 
 /**
@@ -12,22 +12,63 @@ use stdClass;
  * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
-abstract class RequestFactory
+abstract class IncomingRequestFactory
 {
+    /**
+     * Create a request from the supplied superglobal values.
+     *
+     * If any argument is not supplied, the corresponding superglobal value will
+     * be used.
+     *
+     * The IncomingRequest created is then passed to the fromServer() method in
+     * order to marshal the request URI and headers.
+     *
+     * @see fromServer()
+     * @param array $server $_SERVER superglobal
+     * @param array $query $_GET superglobal
+     * @param array $body $_POST superglobal
+     * @param array $cookies $_COOKIE superglobal
+     * @param array $files $_FILES superglobal
+     * @return IncomingRequest
+     */
+    public static function fromGlobals(
+        array $server = null,
+        array $query = null,
+        array $body = null,
+        array $cookies = null,
+        array $files = null
+    ) {
+        $server  = $server  ?: $_SERVER;
+        $query   = $query   ?: $_GET;
+        $body    = $body    ?: $_POST;
+        $cookies = $cookies ?: $_COOKIE;
+        $files   = $files   ?: $_FILES;
+        $request = new IncomingRequest(
+            'php://input',
+            $cookies,
+            [],
+            $query,
+            $body,
+            $files
+        );
+        return self::fromServer($server, $request);
+    }
+
     /**
      * Populates a request object from the given $_SERVER array
      *
      * @param array $server
-     * @param RequestInterface $request
-     * @return RequestInterface The $request provided, only populated with values
+     * @param IncomingRequestInterface $request
+     * @return IncomingRequestInterface The $request provided, only populated with values
      */
-    public static function fromServer(array $server, RequestInterface $request = null)
+    public static function fromServer(array $server, IncomingRequestInterface $request = null)
     {
         $server = self::normalizeServer($server);
 
         if (! $request) {
             $protocol = self::get('SERVER_PROTOCOL', $server, '1.1');
-            $request  = new Request($protocol);
+            $request  = new IncomingRequest();
+            $request->setProtocolVersion($protocol);
         }
 
         $request->setMethod(self::get('REQUEST_METHOD', $server, 'GET'));
@@ -120,10 +161,10 @@ abstract class RequestFactory
      * Marshal the URI from the $_SERVER array and headers
      *
      * @param array $server
-     * @param RequestInterface $request
+     * @param IncomingRequestInterface $request
      * @return Uri
      */
-    public static function marshalUri(array $server, RequestInterface $request)
+    public static function marshalUri(array $server, IncomingRequestInterface $request)
     {
         // URI scheme
         $scheme = 'http';
@@ -163,10 +204,10 @@ abstract class RequestFactory
      * Marshal the host and port from HTTP headers and/or the PHP environment
      *
      * @param array $server
-     * @param RequestInterface $request
+     * @param IncomingRequestInterface $request
      * @return array Array with two members, host and port, at indices 0 and 1, respectively
      */
-    public static function marshalHostAndPort(stdClass $accumulator, array $server, RequestInterface $request)
+    public static function marshalHostAndPort(stdClass $accumulator, array $server, IncomingRequestInterface $request)
     {
         if ($request->hasHeader('host')) {
             return self::marshalHostAndPortFromHeader($accumulator, $request);
@@ -257,10 +298,10 @@ abstract class RequestFactory
      * Marshal the host and port from the request header
      *
      * @param stdClass $accumulator
-     * @param RequestInterface $request
+     * @param IncomingRequestInterface $request
      * @return void
      */
-    private static function marshalHostAndPortFromHeader(stdClass $accumulator, RequestInterface $request)
+    private static function marshalHostAndPortFromHeader(stdClass $accumulator, IncomingRequestInterface $request)
     {
         $accumulator->host = $request->getHeader('host');
         $accumulator->port = null;